+++ contest-02/prob-skiing2/solve.cpp
@@ -1,6 +1,6 @@
 #include <bits/stdc++.h>
 
-#define rep(a, b) for (int a = 0; a < (b); ++a)
+#define rep(a, b) for (ll a = 0; a < (b); ++a)
 #define all(a) (a).begin(), (a).end()
 
 using namespace std;
